Update application/libraries/Datatables.php
amendment to the exec_replace method. 
When using the callback function from a controller, 
it could not see the method in the class. 
The amendment above takes that into account and checks for both the 
function and the class method. By default any function will use a class method if both it and an external function exist. To change this just alter line 438